+++ src/i18n/locales/ru/editor.json
@@ -44,6 +44,25 @@
 		"permissionDenied": "Разрешение на запись запрещено. Пожалуйста, разрешите запись экрана.",
 		"accessibilityAllowAndRetry": "Разрешите OpenScreen доступ к Универсальному доступу, затем снова нажмите запись, чтобы начать обратный отсчет."
 	},
+	"autoCaptions": {
+		"button": "Автосубтитры",
+		"dialogTitle": "Автосубтитры",
+		"dialogDescription": "Выберите, сколько примерно слов показывать в одном субтитре. Время распределяется между словами фразы.",
+		"minWords": "Минимум слов в субтитре",
+		"maxWords": "Максимум слов в субтитре",
+		"wordsCount": "{{count}} слов",
+		"generate": "Создать",
+		"dialogCancel": "Отмена",
+		"generating": "Создание субтитров из звука…",
+		"loadingModel": "Загрузка речевой модели (при первом запуске скачивается ~75 МБ)…",
+		"transcribing": "Распознавание речи…",
+		"busy": "Создание субтитров уже выполняется.",
+		"done": "Добавлено субтитров: {{count}}.",
+		"noneHeard": "Речь не обнаружена.",
+		"noAudio": "В этом видео нет звука, пригодного для расшифровки.",
+		"failed": "Не удалось создать субтитры.",
+		"truncated": "Расшифрованы только первые {{minutes}} мин."
+	},
 	"emptyState": {
 		"title": "Нет открытых проектов",
 		"description": "Импортируйте видео для начала редактирования или загрузите существующий проект OpenScreen.",
