+++ src/i18n/locales/vi/editor.json
@@ -44,6 +44,25 @@
 		"permissionDenied": "Quyền ghi hình bị từ chối. Vui lòng cho phép ghi màn hình.",
 		"accessibilityAllowAndRetry": "Cho phép OpenScreen truy cập Trợ năng, sau đó nhấn ghi lại để bắt đầu đếm ngược."
 	},
+	"autoCaptions": {
+		"button": "Phụ đề tự động",
+		"dialogTitle": "Phụ đề tự động",
+		"dialogDescription": "Chọn khoảng bao nhiêu từ mỗi phụ đề hiển thị cùng lúc. Thời gian được phân bổ cho các từ trong cụm từ đó.",
+		"minWords": "Số từ tối thiểu mỗi phụ đề",
+		"maxWords": "Số từ tối đa mỗi phụ đề",
+		"wordsCount": "{{count}} từ",
+		"generate": "Tạo",
+		"dialogCancel": "Hủy",
+		"generating": "Đang tạo phụ đề từ âm thanh…",
+		"loadingModel": "Đang tải mô hình giọng nói (lần đầu sử dụng sẽ tải ~75 MB)…",
+		"transcribing": "Đang chuyển lời nói thành văn bản…",
+		"busy": "Việc tạo phụ đề đang được tiến hành.",
+		"done": "Đã thêm {{count}} phụ đề.",
+		"noneHeard": "Không phát hiện thấy lời nói.",
+		"noAudio": "Video này không có âm thanh dùng được để chuyển thành văn bản.",
+		"failed": "Không thể tạo phụ đề.",
+		"truncated": "Chỉ {{minutes}} phút đầu tiên được chuyển thành văn bản."
+	},
 	"emptyState": {
 		"title": "Không có dự án nào được mở",
 		"description": "Nhập video để bắt đầu chỉnh sửa hoặc tải một dự án OpenScreen hiện có.",
